services/python/logic/requirements.txt
numpy
matplotlib
flask
requests
flask-sqlalchemy
pymysql
pillow
